# optimize memory allocs during reconstruct