… pkg/bpool/bpool.go
… pkg/bpool/bpool_test.go
… vendor/github.com/klauspost/reedsolomon/reedsolomon.go
… vendor/vendor.json