@aead 2017-09-26
Here quick benchstats of a 2x2 setup 2 disks down and 100MB object
```
name        old time/op    new time/op    delta
ReadFile-4     180ms ± 0%     164ms ± 0%   ~     (p=1.000 n=1+1)

name        old alloc/op   new alloc/op   delta
ReadFile-4     109MB ± 0%      22MB ± 0%   ~     (p=1.000 n=1+1)

name        old allocs/op  new allocs/op  delta
ReadFile-4       433 ± 0%       415 ± 0%   ~     (p=1.000 n=1+1)
```