@donatello 2017-09-27
@aead You don't seem to be following the line length rule (at least in changed lines) that you wanted to have. 100 columns? We may bring up this point in the next meeting again, as some were not present last time. In addition, need to decide if review is the right place to point out this issue.