@fwessels 2017-09-27
@donatello We can discuss this again, but let's also please separate main causes and side effects. Code looks perfectly fine (and improved).