@aead 2017-09-27
@donatello Your right about the length / column issue.
The problem is what do we do with e.g. [this](https://github.com/minio/minio/pull/4964/files#diff-cc000e8f9417c174d87c19bd6bf030d1R26). If I understood the discussion here correctly I should open an issue for refactoring it. Just blindly wrapping the line after 100CL was considered as not a sufficient approach (which I agree). So I open an issue for this.